requirements.txt
Django==3.1.5
djangorestframework==3.12.2
django-rest-auth[with_social]==0.9.5
six==1.15.0
psycopg2==2.8.6
python==3.9.2
python-decouple==3.4
dj-database-url==0.5.0
Pillow==8.1.2
django-filter==2.4.0

django-auto-prefetching==0.1.9